refactor : Answer 패키지 생성 및 관련 클래스 파일 이동 - Controller 파일에 UserDetailsImpl 클래스 추가
