+++ src/main/java/com/example/team258/answer/AnswerController.java
@@ -1,14 +1,11 @@
-package com.example.team258.controller;
+package com.example.team258.answer;
 
-import com.example.team258.dto.AnswerRequestDto;
-import com.example.team258.dto.AnswerResponseDto;
-import com.example.team258.service.AnswerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-
+import com.example.team258.security.UserDetailsImpl;
 import java.util.List;
 
 @Controller
