+++ src/main/java/com/example/team258/entity/Survey.java
@@ -1,5 +1,6 @@
 package com.example.team258.entity;
 
+import com.example.team258.answer.Answer;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
