+++ src/main/java/com/example/team258/repository/AnswerRepository.java
@@ -1,6 +1,6 @@
 package com.example.team258.repository;
 
-import com.example.team258.entity.Answer;
+import com.example.team258.answer.Answer;
 import com.example.team258.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
